# backend/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: MTB REST API
  version: 1.0.0
  description: REST API of the MTB service.
...
servers:
  - url: /api
paths:
...
  /episodes/{episodeId}/rebiopsyrequests:

    post:
      summary: add rebiopsy request for an episode.
      description: add rebiopsy request for an episode.
      operationId: addRebiopsyRequest
      tags:
        - rebiopsy_request
      parameters:
        - name: episodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the episode
      requestBody:
        description: add rebiopsy request for an episode.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RebiopsyRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RebiopsyRequest'
        400:
          description: Foreign key constraint violated (Check episodeId and specimen in your request)
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
        405:
          description: Invalid Input (The given episode Id in the body is not the same as in the path)
    get:
      summary: List all rebiopsy request for an episode.
      operationId: getAllRebiopsyRequests
      tags:
        - rebiopsy_request
      parameters:
        - name: episodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the episode
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RebiopsyRequest'
        400:
          description: Foreign key constraint violated (Episode not found)
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
  /episodes/{episodeId}/rebiopsyrequests/{rebiopsyRequestId}:

    get:
      summary: get rebiopsy request for a specific rebiopsyRequestId.
      operationId: getRebiopsyRequest
      tags:
        - rebiopsy_request
      parameters:
        - name: episodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the episode
        - name: rebiopsyRequestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the rebiopsy request
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RebiopsyRequest'
        400:
          description: Foreign key constraint violated (Episode not found or is not the same as the saved in the database)
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
        404:
          description: rebiopsy request not found

    put:
      summary: Change rebiopsy request for an episode.
      description: Change rebiopsy request for an episode.
      operationId: updateRebiopsyRequest
      tags:
        - rebiopsy_request
      parameters:
        - name: episodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the episode
        - name: rebiopsyRequestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the rebiopsy request
      requestBody:
        description: Change rebiopsy request for an episode.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RebiopsyRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RebiopsyRequest'
        400:
          description: Foreign key constraint violated (Check EpisodeId and Specimen)
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
        404:
          description: rebiopsy request not found
        405:
          description: Invalid Input

    delete:
      summary: Delete rebiopsy request for an episode.
      description: Delete rebiopsy request for an episode.
      operationId: deleteRebiopsyRequest
      tags:
        - rebiopsy_request
      parameters:
        - name: episodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of the episode
        - name: rebiopsyRequestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The id of rebiopsy request
      responses:
        200:
          description: OK
        401:
          description: Unauthorized - The provided credentials are unknown
        403:
          description: Forbidden
        404:
          description: rebiopsy request not found
...
components:
  schemas:
...
    RebiopsyRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        episodeId:
          type: string
          format: uuid
        specimen:
          type: string
          format: uuid
        issuedOn:
          type: string
          format: date
...